Ulbra AA
Drops Em Comentario para Nao haver a repetiçao de itens na lista

+++ UlbraAA(C)/DB/CursoRepositorio.cs
@@ -38,7 +38,10 @@ namespace UlbraAA_C_.DB.Classes
        public static void DropCurso() 
        {
            database db = getDataBase();
-           db.DeleteDatabase();
+           if (db.DatabaseExists() == true) 
+           {
+               db.DeleteDatabase();
+           }
            
        }
